HP 제거
