+++ src/main/resources/application.yml
@@ -4,6 +4,7 @@ spring:
     url: jdbc:h2:mem:mingledb
     username: sa
     password:
+
   jpa:  # https://joyerim.tistory.com/6#%E2%96%AA%C2%A0JPA-1
     defer-datasource-initialization: true
     show-sql: true
@@ -21,3 +22,7 @@ spring:
     live:
       atk: 300000
       rtk: 3000000
+
+  h2:
+    console:
+      enabled: true
